Minor map tweaks

+++ app/src/main/cpp/jpeg2000decoder.c
@@ -60,7 +60,7 @@ static OPJ_BOOL mem_seek_fn(OPJ_OFF_T p_nb_bytes, void* p_user_data) {
 
 JNIEXPORT jobject JNICALL
 Java_com_dbf_aqhi_jpeg_Jpeg2000Decoder_decodeJpeg2000(
-        JNIEnv *env, jclass clazz, jbyteArray data, jint offset, jint length, jint scale, jint max_alpha) {
+        JNIEnv *env, jclass clazz, jbyteArray data, jint offset, jint length, jfloat scale, jint max_alpha) {
 
     //Get pointer to array elements to avoid copying the entire array
     jbyte* all_data = (*env)->GetPrimitiveArrayCritical(env, data, NULL);
@@ -159,7 +159,7 @@ Java_com_dbf_aqhi_jpeg_Jpeg2000Decoder_decodeJpeg2000(
     jbyte* outPixels = (jbyte*)(*env)->GetByteArrayElements(env, pixelArray, NULL);
 
     const int range = (max_val - min_val);
-    const int scaleFactor = 255 * scale;
+    const float scaleFactor = 255 * scale;
     //Use first component only (greyscale)
     for (int i = 0; i < pixelCount; ++i) {
         int grey = ((comp->data[i] - min_val) * scaleFactor) / range;
